Consolidation: test_miso.py
